use repodata_from_packages.json if exists
This file includes yanked packages, so we use it to do the mirror,
although we upload and honor the current repodata.json. In addition,
I am adding a helper script to run the registry so the developer
does not need to remember / copy paste from the readme, and support
for running in serial to the other mirror functions to be consistent!

Signed-off-by: vsoch <vsoch@users.noreply.github.com>